Adjusted tests, added Testcases to 214
