+++ tests/protocols/gmpv208/testcmds/__init__.py
@@ -136,9 +136,6 @@ from .test_get_groups import GmpGetGroupsTestCase
 
 from .test_get_note import GmpGetNoteTestCase
 from .test_get_notes import GmpGetNotesTestCase
-from .test_get_scan_config_nvt import GmpGetScanConfigNvtTestCase
-from .test_get_nvt_families import GmpGetNvtFamiliesTestCase
-from .test_get_scan_config_nvts import GmpGetScanConfigNvtsTestCase
 from .test_get_override import GmpGetOverrideTestCase
 from .test_get_overrides import GmpGetOverridesTestCase
 from .test_get_permission import GmpGetPermissionTestCase
